+++ src/io/flutter/run/common/TestLineMarkerContributor.java
@@ -68,26 +68,9 @@ public abstract class TestLineMarkerContributor extends RunLineMarkerContributor
             return new RunLineMarkerContributor.Info(icon, ExecutorAction.getActions(), tooltipProvider);
           }
         }
-        else if (dartId.getParent().getParent() instanceof DartFunctionDeclarationWithBodyOrNative) {
-          if (testConfigUtils instanceof TestConfigUtils) {
-            // TODO(messick) Find a better way to eliminate duplicate pop-up menu entries.
-            // The issue is that there are two contributors, one for normal Flutter, one for Bazel,
-            // and they should not both produce contributions at the same time.
-            return null;
-          }
-          if ("main".equals(dartId.getText())) {
-            // There seems to be an intermittent timing issue that causes the first test call to not get marked.
-            // Priming the cache here solves it.
-            testConfigUtils.refreshOutline(element);
-            TestType testCall = TestType.MAIN;
-            final Icon icon = getTestStateIcon(element, testCall.getIcon());
-            final Function<PsiElement, String> tooltipProvider =
-              psiElement -> testCall.getTooltip(psiElement, testConfigUtils);
-            return new RunLineMarkerContributor.Info(icon, ExecutorAction.getActions(), tooltipProvider);
-          }
-        }
       }
     }
+
 
     return null;
   }