+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### Removed
 - Support for platform version 2025.1.
+- Bazel run and test configurations.
 
 ### Fixed